Move thread_pool implementaion to concurrency folder
